[fix] 03/12 파일 경로 수정
