+++ app/(main)/user/funding/[novelId]/FundingIntro.styled.tsx
