+++ app/(main)/user/funding/[novelId]/create/StartFunding.styled.tsx
